+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file. This
 project adheres to [Semantic Versioning](http://semver.org/).
 
+## (next)
+- CHANGED: to always ICMP-reject connection according to RFC3360
+- FIXED: to accept unmatched TCP-RST on interfaces with REJECT policy
+
 ## 1.0.0 (2018-03-26)
 - FIXED: minor Ruby warnings